flask-app/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
python-dotenv==1.0.0
bcrypt==4.0.1
APScheduler==3.10.4
Werkzeug==2.3.7
SQLAlchemy==2.0.23
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
email-validator==2.1.0 